Touching up wording.

+++ LulCaster.UI.WPF/Pages/WireFramePage.xaml.cs
@@ -26,7 +26,7 @@ namespace LulCaster.UI.WPF.Pages
   {
     #region "Private Members"
     private readonly ScreenCaptureTimer _screenCaptureTimer;
-    private const double CAPTURE_TIMER_INTERVAL = 1000;
+    private const int CAPTURE_TIMER_INTERVAL = 1000;
     private readonly BoundingBoxBrush _boundingBoxBrush = new BoundingBoxBrush();
     private readonly Dictionary<string, Rectangle> _boundingBoxCollection = new Dictionary<string, Rectangle>(); //TODO: This will live in the region configuration tool
     private Rectangle _currentBoundingBox; //TODO: This will live in the region configuration tool
